# public/openapi.yaml
openapi: 3.1.0
info:
  title: Kyozo Pro API
  version: "1.0.0"
  summary: Public API for the Kyozo community platform.
  description: |
    The Kyozo Pro API exposes the platform's email, WhatsApp, AI, broadcast, community, and waitlist
    surfaces for external automation and MCP-driven agents.

    All endpoints are versioned under `/api/v1/`. Authentication is by API key in the
    `x-api-key` header (or `Authorization: ApiKey <key>`). Mint keys with
    `pnpm mint-api-key --name "..." --scopes "..."`.
  contact:
    name: Kyozo
    url: https://kyozo.com
  license:
    name: Proprietary
servers:
  - url: https://pro.kyozo.com
    description: Production
  - url: http://localhost:9003
    description: Local dev
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/email/setup-domain:
    get:
      tags: [Email]
      summary: Get sending-domain status
      operationId: getEmailDomain
      parameters:
        - in: query
          name: domain
          schema: { type: string }
          description: FQDN to check (e.g. `acme.kyozo.com`).
      responses:
        "200":
          description: Domain status (or `not_found`).
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain: { type: string }
                  status: { type: string }
                  domainId: { type: string }
                  records:
                    type: array
                    items: { $ref: "#/components/schemas/DnsRecord" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
    post:
      tags: [Email]
      summary: Add or verify a sending domain
      operationId: setupEmailDomain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                handle:
                  type: string
                  description: Community handle. Implies domain `<handle>.kyozo.com` if `domain` not given.
                domain:
                  type: string
                  description: Full FQDN to set up (overrides `handle`).
                action:
                  type: string
                  enum: [list, get, verify, add]
                  default: add
            examples:
              addByHandle:
                value: { handle: acme }
              verifyExisting:
                value: { domain: acme.kyozo.com, action: verify }
      responses:
        "200":
          description: Setup or status result.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }
  /api/v1/whatsapp/send-template:
    post:
      tags: [WhatsApp]
      summary: Send a WhatsApp template message
      operationId: sendWhatsappTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendWhatsappTemplateRequest"
            examples:
              simple:
                summary: Send sample_1 template with no parameters
                value:
                  to: "+15551234567"
                  template:
                    name: sample_1
                    language: en_US
              withBodyParams:
                summary: Body with two text parameters
                value:
                  to: "+15551234567"
                  template:
                    name: appointment_reminder
                    language: { code: en_US }
                    components:
                      - type: BODY
                        parameters:
                          - { type: text, text: "Ashok" }
                          - { type: text, text: "Tuesday" }
      responses:
        "200":
          description: Provider accepted the message.
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502":
          description: Upstream provider rejected the request.
        "500": { $ref: "#/components/responses/ServerError" }
  /api/v1/whatsapp/templates:
    get:
      tags: [WhatsApp]
      summary: List available WhatsApp templates
      operationId: listWhatsappTemplates
      responses:
        "200":
          description: List of templates available on the connected 360dialog account.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  templates:
                    type: array
                    items: { $ref: "#/components/schemas/WhatsappTemplate" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502":
          description: Upstream provider error.
...
  /api/v1/waitlist/join:
    post:
      tags: [Waitlist]
      summary: Submit a waitlist (access request) entry
      operationId: waitlistJoin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WaitlistJoinRequest"
      responses:
        "200":
          description: Submitted (or already submitted/approved).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  requestId: { type: string }
                  status: { type: string, enum: [pending, approved, rejected] }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Issue keys with `pnpm mint-api-key`. Each key has a scope list; routes
        require specific scopes (e.g. `email:send`, `whatsapp:send`).
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, or revoked API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServerError:
      description: Server error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
        details: {}
...
    DnsRecord:
      type: object
      properties:
        record: { type: string }
        name: { type: string }
        type: { type: string }
        ttl: { type: string }
        status: { type: string }
        value: { type: string }
        priority: { type: integer }
    SendWhatsappTemplateRequest:
      type: object
      required: [to, template]
      properties:
        to:
          type: string
          description: Recipient phone number in E.164 format.
        wabaAccountId:
          type: string
          description: Optional 360dialog WABA account id (sets `D360-WABA-ACCOUNT-ID`).
        template:
          type: object
          required: [name]
          properties:
            name: { type: string }
            language:
              oneOf:
                - { type: string }
                - {
                    type: object,
                    properties: { code: { type: string } },
                  }
            components:
              type: array
              items:
                type: object
                properties:
                  type: { type: string, enum: [HEADER, BODY, FOOTER, BUTTON] }
                  text: { type: string }
                  parameters:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          { type: string, enum: [text, image, document, video] }
                        text: { type: string }
                        image:
                          type: object
                          properties: { link: { type: string, format: uri } }
                        document:
                          type: object
                          properties: { link: { type: string, format: uri } }
                        video:
                          type: object
                          properties: { link: { type: string, format: uri } }
    WhatsappTemplate:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        language: { type: string }
        components:
          type: array
          items:
            type: object
            additionalProperties: true
...
    WaitlistJoinRequest:
      type: object
      required: [email, firstName, lastName, phone, agreedToPrivacy]
      properties:
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
        newsletter: { type: boolean }
        whatsapp: { type: boolean }
        agreedToPrivacy: { type: boolean }
        source: { type: string }